+++ rpcs3/Emu/Cell/PPUThread.h
@@ -59,6 +59,54 @@ struct ppu_thread_params
 	u64 arg1;
 };
 
+struct cmd64
+{
+	u64 m_data = 0;
+
+	constexpr cmd64() noexcept = default;
+
+	struct pair_t
+	{
+		u32 arg1;
+		u32 arg2;
+	};
+
+	template <typename T, typename T2 = simple_t<T>>
+	cmd64(const T& value)
+		: m_data(std::bit_cast<u64, T2>(value))
+	{
+	}
+
+	template <typename T1, typename T2>
+	cmd64(const T1& arg1, const T2& arg2)
+		: cmd64(pair_t{std::bit_cast<u32>(arg1), std::bit_cast<u32>(arg2)})
+	{
+	}
+
+	explicit operator bool() const
+	{
+		return m_data != 0;
+	}
+
+	template <typename T>
+	T as() const
+	{
+		return std::bit_cast<T>(m_data);
+	}
+
+	template <typename T>
+	T arg1() const
+	{
+		return std::bit_cast<T>(std::bit_cast<pair_t>(m_data).arg1);
+	}
+
+	template <typename T>
+	T arg2() const
+	{
+		return std::bit_cast<T>(std::bit_cast<pair_t>(m_data).arg2);
+	}
+};
+
 class ppu_thread : public cpu_thread
 {
 public:
